Update : Drops, Items, Enemies, and some more fixes!
Un update ce aduce in sfarsit putin gameplay.

ADDED:
+ Inamici de tip "Manta Ray" la fiecare nivel (Temporar)
+ Health Bar pentru inamici
+ Shield Bar pentru inamici
+ Item drops
+ Damage to enemies

FIXED:
= Enemy drops
= Enemy shield and health

+++ Project/Gravity Shift/Assets/Scripts/HullItem.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class HullItem : EquippableItem
 {
-    public int maxHullHP = 0;
+    public float maxHullHP = 0;
     public float forwardSpeed = 0;
     public float rotationSpeed = 0;
