+++ Project/Gravity Shift/Assets/Scripts/ReactorItem.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu]
 public class ReactorItem : EquippableItem {
 
-    public int maxEnergy = 0;
-    public int rechargeRate = 0;
-    public int timeRecharge = 0;
+    public float maxEnergy = 0;
+    public float rechargeRate = 0;
+    public float timeRecharge = 0;
 
 }
